Don't scan through entire block for all intra-block seeks.
Previously, when seeking further in the same block, libhdfs3 would
scan all data content between current offset and to-be-seeked offset
to avoid issuing a new RPC call for short seeks. But an intra-block
seek can still be a substantial amount of data and require many
requests.

The native java client uses a more optimal method: it only scans when
the to-be-seeked offset lies in the socket buffer; otherwise it will
send a new OpReadBlockProto RPC call. Update libhdfs3 to use this
superior method.

From original patch by Van Zhong.